+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * `KYCRecovery` reviewable request ingestion
 * `GetKYCRecoveryRequests` to handle `/v3/kyc_recovery_requests`, `/v3/kyc_recovery_requests/{id}`
 * Docs on kyc recovery
+* `/v3/accounts/{id}/sales/{sale_id}` endpoint to get sale by id, if account is allowed to participate
+* `GetSaleForAccount` to handle `/v3/accounts/{id}/sales/{sale_id}`
 
 ### Fixed